Separate DashGridLayoutView into separate component from DashView
