+++ desktop/cmp/dash/impl/DashView.js
@@ -6,14 +6,8 @@
  */
 import {frame} from '@xh/hoist/cmp/layout';
 import {hoistCmp, ModelPublishMode, refreshContextView, RenderMode, uses} from '@xh/hoist/core';
-import {contextMenu} from '@xh/hoist/desktop/cmp/contextmenu';
 import {elementFromContent} from '@xh/hoist/utils/react';
 import {useRef} from 'react';
-import {Icon} from '../../../../icon';
-import {DashGridLayoutContainerModel} from '../../dashGrid';
-import {panel} from '../../panel';
-import {button} from '../../button';
-import {popover, Position} from '@xh/hoist/kit/blueprint';
 import {DashViewModel} from '../DashViewModel';
 
 /**
@@ -49,55 +43,12 @@ export const dashView = hoistCmp.factory({
         ) {
             return null;
         }
-
-        if (containerModel instanceof DashGridLayoutContainerModel) {
-            return panel({
-                className,
-                compactHeader: true,
-                title: model.title,
-                icon: model.icon,
-                headerItems: [
-                    popover({
-                        position: Position.BOTTOM,
-                        minimal: true,
-                        target: button({
-                            icon: Icon.ellipsisVertical()
-                        }),
-                        content: contextMenu({
-                            menuItems: [
-                                {
-                                    text: 'Rename',
-                                    icon: Icon.edit(),
-                                    disabled: !viewSpec.allowRename,
-                                    actionFn: () => containerModel.renameView(id)
-                                },
-                                {
-                                    text: 'Duplicate',
-                                    icon: Icon.copy(),
-                                    actionFn: () =>
-                                        containerModel.addView(viewSpec.id, model.positionParams)
-                                },
-                                '-',
-                                {
-                                    text: 'Remove',
-                                    icon: Icon.cross(),
-                                    disabled: !viewSpec.allowRemove,
-                                    actionFn: () => containerModel.removeView(id)
-                                }
-                            ]
-                        })
-                    })
-                ],
-                item: elementFromContent(viewSpec.content, {flex: 1, viewModel: model})
-            });
-        } else {
-            return frame({
-                className,
-                item: refreshContextView({
-                    model: refreshContextModel,
-                    item: elementFromContent(viewSpec.content, {flex: 1})
-                })
-            });
-        }
+        return frame({
+            className,
+            item: refreshContextView({
+                model: refreshContextModel,
+                item: elementFromContent(viewSpec.content, {flex: 1})
+            })
+        });
     }
 });
